+++ nn_dataflow/__init__.py
@@ -21,6 +21,7 @@ program. If not, see <https://opensource.org/licenses/BSD-3-Clause>.
 from . import DataCategoryEnum
 from . import LoopBlocking
 from . import LoopBlockingSolver
+from . import LoopEnum
 from . import MemHierEnum
 from . import ParallelEnum
 from . import Partition
@@ -35,14 +36,15 @@ from .LoopBlockingScheme import LoopBlockingScheme
 from .MapStrategy import MapStrategy, MapStrategyEyeriss
 from .NestedLoopDesc import NestedLoopDesc
 from .Network import Network
+from .NodeRegion import NodeRegion
 from .NNDataflowScheme import NNDataflowScheme
 from .Option import Option
 from .PartitionScheme import PartitionScheme
 from .PhyDim2 import PhyDim2
-from .Resource import NodeRegion, Resource
+from .Resource import Resource
 from .Scheduling import SchedulingCondition, SchedulingResult, Scheduling
 
 from .NNDataflow import NNDataflow
 
-__version__ = '1.5-rc.5'
+__version__ = '1.5-rc.6'