+++ tools/eyeriss_search.py
@@ -98,19 +98,27 @@ def do_scheduling(args):
     size_gbuf = args.gbuf / word
     size_regf = args.regf / word
 
+    proc_region = NodeRegion(dim=dim_nodes,
+                             origin=PhyDim2(0, 0),
+                             type=NodeRegion.PROC)
+
     if args.mem_type == '2D':
         # Memory nodes are on two sides.
-        mem_regions = (NodeRegion(dim=PhyDim2(h=dim_nodes.h, w=1),
-                                  origin=PhyDim2(h=0, w=0)),
-                       NodeRegion(dim=PhyDim2(h=dim_nodes.h, w=1),
-                                  origin=PhyDim2(h=0, w=dim_nodes.w - 1)))
+        data_regions = (NodeRegion(dim=PhyDim2(h=dim_nodes.h, w=1),
+                                   origin=PhyDim2(h=0, w=0),
+                                   type=NodeRegion.DATA),
+                        NodeRegion(dim=PhyDim2(h=dim_nodes.h, w=1),
+                                   origin=PhyDim2(h=0, w=dim_nodes.w - 1),
+                                   type=NodeRegion.DATA))
     elif args.mem_type == '3D':
         # All nodes have memory.
-        mem_regions = (NodeRegion(dim=dim_nodes, origin=PhyDim2(0, 0)),)
+        data_regions = (NodeRegion(dim=dim_nodes,
+                                   origin=PhyDim2(0, 0),
+                                   type=NodeRegion.DATA),)
 
-    resource = Resource(dim_nodes=dim_nodes,
+    resource = Resource(proc_region=proc_region,
+                        data_regions=data_regions,
                         dim_array=dim_array,
-                        mem_regions=mem_regions,
                         size_gbuf=size_gbuf,
                         size_regf=size_regf)